Start django-mailer's sending functions when we send emails
Django-mailer requires firing these functions in some way,
otherwise the Message objects will just be stored in the DB
and never sent.

This way of starting django-mailer is a good addition to and
should be used together with cron jobs suggested in its docs.

NOTE: We do not attempt to purge mail logs here. This can be
done with `purge_mail_log` management command, which should
ideally be registered as a cron job as suggested in the docs.