Localize decimal fields for reservations' endpoint
This way, when API's language is set to Finnish or any other
language, where commas are the standard decimal separator, the
endpoint will accept both dots and commas in decimal fields.

NOTE: if the language is set to languages, where dot is the
standard decimal separator, f.ex. English, then only dots will
be allowed.