Update harbors fixtures
Some of the Harbor model fields that were nullable now are not.
Replacing their null values with blank strings to make fixtures
installable again.

Fixtures are mainly needed for dev purposes, we installed them
in production and test instance of the backend once, after which
they need not be updated.