Update script that parses berths data from the TIMMI csv file
We want to have info about single berths now: their dimensions
and mooring types.

Data about harbors and piers will be fetched from the currently
existing "harbors" app (which will be eventually deleted from
the project). To do this, all we need to know is servicemap
ids of harbors and identifiers of piers.