Add new resources' models
We want to store data about single berths and winter storage
places, so that they can be later offered to customers.

During the transition period, we will have the old 'harbors' app
and will use some of the data from it to populate the models for
the new 'resources' app.