Add mngmnt command that imports harbors, piers and berths
This management command is used to import harbors, piers and
berths from a JSON that was parsed by scripts/parse_harbors_csv.py
script from a CSV file exported from the old TIMMI software of
Helsinki's boating services.

It is probably not going to be needed in the future, as all the
data will be dumped into fixtures that can be installed for new
instances of this project. It might be useful in the future, if
there will be a need to do imports from new csv files exported
from TIMMI (if it, for example, will have more info than the one
that was used in development this time).