Move factory-boy from dev reqs to regular requirements
We are passing in-memory objects, based on our models, to
django-ilmoitin's preview context. That requires factory-boy's
packages to be installed, but some instances only use regular
requirements and skip dev requirements, so they do not get it
installed. If we move factory to reqular reqs, we will ensure
that it will be installed there too.