Rename some mngmt commands in harbors app
We have commands with the same names in the new `resources` app.
Since `harbors` app is defined before `resources` in the installed
apps, they become overriden by these commands, so to avoid that
let's rename them by prefixing the app label. In the future, the
harbors app will be removed altogether.