Delete old GraphQLDjangoTestClient
We used it to, because it allowed to pass Accept-Language header
with the request, as it went through Django's LocaleMiddleware,
which switches the language based in the header value.
Now we use Graphene's test Client for tests and simply enforce
English in our API calls. It is more flexible and would make
certain tests easier (e.g. file upload tests).

Django already has tests for LocaleMiddleware:
https://github.com/django/django/blob/2.2.11/tests/i18n/tests.py#L1244
so we can just skip testing that Accept-Language header works.