Check that lease dates don't overlap
By only checking that the end date was earlier than the lease we were
trying to add limited to only adding leases "in the future", or after
that leases' end.

Now, checking both for start and end date, we ensure that new leases can
be created also for past leases and that there's no overlap between
them.