Add an OrderToken model to handle payment tokens
When a payment attempt is made, instead of directly calling Bambora,
we try to fetch the token associated with that order number and the
created_at timestamp. When that token has expired (midnight on the sixth
day after it was created), a new order is submitted to Bambora and the
new token is stored.