Set the correct lease status on creation
When creating the lease, we are just copying the previous one which will
be on paid status. For proper bookkeeping, we should set it as drafted
to follow the correct flow from drafted -> offered/error -> paid/rejected.

The lease would in any case be marked as offered/error by the script,
but better to follow the right steps.