Remove automatic creation of order lines
On the previous pricing schema, the fixed additional services
(electricity, gate, lighting, mooring, waste collection, and water) were
billed separately but the order lines were generated automatically when
creating the order, based on the services available on the pier/harbor.

The new pricing schema already includes those services on the berth
prices, so we don't need to bill them separately.