Adjust the lease termination logic
When a lease is terminated, if no end date was provided, adjust
the default end_date to be max(today, start_of_the_lease_season).
This needs to be adjusted for the special case if a lease is
terminated before the season starts, so that the condition
`start_date>=end_date` is fulfilled.

Refs VEN-1179