Add mutation for switching the berth of a berth lease
Add SwitchBerthMutation that allows to switch the berth from a
berth lease to another berth that is not linked to a lease. The
berth lease must have a contract and the status must be paid in
order to switch the berth for it.
In practice the old berth lease is terminated and a new one is
created with the same information as the old one except with the
berth changed to the new provided berth.

Refs VEN-1191