Add a mapping between old and new harbors
The old harbors app was the first version of the whole system. It had a
simple information architecture which covered the MVP needs on the
moment, necessary only to release the applications.

When applications was released and the development started to shift
towards the admin UI, a new, more organised, clearly divided
categorisation came with the resources app. The "harbors" on the old
harbors app were now split into Piers and Harbors.

Some of the old harbors corresponded directly to a new harbor on a 1:1
mapping, some didn't but they mapped to a new pier.

Adding this mapping allows to connect the old harbors (used for
applications) with the new ones (used everywhere else), so some
functionality can be unblocked, like using the berth switch
applications.