Make django admin customer fields use select2
Using Select2 through autocomplete fields removes the overhead of
prefetching all the existing objects, it only prefetches a small subset
and as you start typing, it looks for the ones that match the criteria.

This is especially useful for fields like customer which on the real use
case have over 10K instances and add a big overhead when loading a
single instance.

https://docs.djangoproject.com/en/3.1/ref/contrib/admin/#django.contrib.admin.ModelAdmin.autocomplete_fields