Fix algorithm to determine if berth is available
A few extra considerations had to be taken: for leases which end during
the season, the berth should be available ONLY after the day when the
lease ends (e.g. if the lease ends on 10.7, the berth will not be
available until 10.7, after that, it will be marked as available).