Add GDPR api serializers to models
Add serializers so that the customer can query their personal
data.
I've tried to keep the duplicate information to a mininum so that
for example Boats are listed only once in the customer profile
and when they are linked to Leases, for example, we only show the
boat id in those instances.
There is a bug in the helsinki-profile-gdpr-api that does not
automatically serialize forward facing relations (eg ForeignKeys)
so, for now, those have to be serialized "manually". For example:
`"name": "berth", "accessor": lambda x: x.serialize()`

Refs VEN-1258