Add GDPR API tests
Add test tools & tests for testing the GDPR API and the
serializers.
Had to add the default to the TOKEN_AUTH_AUTHSERVER_URL env
variable because it seemed impossible to get the tests working
without that value. Apparently
`helusers.settings.api_token_auth_settings` are loaded when the
test run is started so I did not figure out how to change the
settings afterwards in the tests.

Refs VEN-1258