Fix resending of failed invoices
When sending the invoices, we only fetch the profiles for the customers
with “renewable” leases. The error leases are not considered “renewable”,
so we’re not fetching the profiles for those customers and things are breaking.

This fix adds the profiles from failed orders to the list of profiles to
be fetched from the Profiili service.

Additionally, the logging is improved to have better visibility of the
process.