fix: add backward compatibility to Django's Choices enums
VEN-1175.

When Berth was upgraded from Django 3.2 to 4.2, the Django's Choices
(`django.db.models.enums.Choices`) are returning the class name instead
of the enum value. Instead of trying to find every single usage of those
Choices fields (which there are pretty many), a ChoicesMixin mixin class
is to provide `__repr__` -function, that brings back the old value
calling representation of Django Choices.